% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorageR.r
\docType{package}
\name{cloudstorageR}
\alias{cloudstorageR}
\alias{cloudstorageR-package}
\title{An R package to interact with Google Cloud Storage.}
\description{
An R package to interact with Google Cloud Storage.
}

