% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.r
\name{delete_obj}
\alias{batch_delete_objs}
\alias{delete_obj}
\title{Delete an object in a bucket.}
\usage{
delete_obj(bucket, obj, token = get_access_cred())

batch_delete_objs(bucket, objs, token = get_access_cred())
}
\arguments{
\item{bucket}{bucket name}

\item{token}{a valid OAuth2.0 token}

\item{objs}{list of object names}

\item{object}{object name}

\item{bucket}{bucket name}

\item{token}{a valid OAuth2.0 token}
}
\value{
TRUE if the object is successfully deleted; FALSE otherwise

response
}
\description{
Delete an object in a bucket.

Batch delete multiple objects
}

