% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r
\name{get_access_cred}
\alias{get_access_cred}
\alias{reset_access_cred}
\alias{set_access_cred}
\title{Get and set access credentials}
\usage{
get_access_cred()

set_access_cred(value)

reset_access_cred()
}
\arguments{
\item{value}{new access credentials, as returned by
\code{\link[httr]{oauth2.0_token}}}
}
\description{
Get and set access credentials
}
\section{API console}{

To manage your google projects, use the API console:
\url{https://console.cloud.google.com/}
}
\seealso{
Scope documentation:
 \url{https://developers.google.com/identity/protocols/googlescopes}

Cloud Storage OAuth documentation:
 \url{https://cloud.google.com/storage/docs/json_api/v1/how-tos/authorizing}
}
\keyword{internal}

