% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.r
\name{simple_upload}
\alias{simple_upload}
\title{A simple upload of a file to a Cloud Storage bucket}
\usage{
simple_upload(file, bucket, name = "uploadedObject", type = NULL,
  token = get_access_cred())
}
\arguments{
\item{file}{Name of the uploaded file.}

\item{bucket}{Bucket name.}

\item{name}{Object name.}

\item{type}{Media type.}

\item{token}{A valid OAuth2.0 token.}
}
\description{
A simple upload of a file to a Cloud Storage bucket
}

