% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.r
\name{upload_df_as_csv}
\alias{upload_df_as_csv}
\title{Upload a data frame as a CSV file to Cloud Storage}
\usage{
upload_df_as_csv(df, bucket, name = deparse(substitute(df)),
  token = get_access_cred())
}
\arguments{
\item{df}{data frame name}

\item{bucket}{bucket name}

\item{name}{object name}

\item{token}{a valid OAuth token}
}
\value{
uploaded object's self link
}
\description{
Upload a data frame as a CSV file to Cloud Storage
}

