% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.r
\name{upload_file_to_cs}
\alias{upload_file_to_cs}
\title{Upload a file to a Cloud Storage bucket}
\usage{
upload_file_to_cs(file, bucket, name = "uploadedObject", type = NULL,
  token = get_access_cred())
}
\arguments{
\item{file}{Name of the uploaded file}

\item{bucket}{Bucket name}

\item{name}{Object name}

\item{type}{Media type}
}
\value{
uploaded object's self link
}
\description{
Upload a file to a Cloud Storage bucket
}

