% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.r
\name{delete_obj}
\alias{delete_obj}
\title{Delete an object in a bucket.}
\usage{
delete_obj(bucket, object, token = get_access_cred())
}
\arguments{
\item{bucket}{bucket name}

\item{object}{object name}

\item{token}{a valid OAuth2.0 token}
}
\value{
TRUE if the object is successfully deleted; FALSE otherwise
}
\description{
Delete an object in a bucket.
}

